Adjust how to check for keyswap commands
Since I make keyswap commands using `lambda' I was checking for a
possible one of these commands using `listp'.  This was being used when
*undoing* having made some swaps.

After an update to emacs 30 it seems that the `lambda' form was instead
returning an `interpreted-function'.  This meant that when undoing the
keyswap I would not actually recognise my mappings and would hence not
remove the wrappers.

This wasn't causing much of a problem in most buffers since I would not
turn off/on keyswap mode very often.  However with `isearch-mode' it
would add another layer of wrapping each time I entered it.  This
eventually drastically lowered performance.

Here we adjust our check for emacs versions greater than 30 so that we
correctly identify our keyswap functions.