Make an attempt to avoid a compilation warning
I want to define an action if another plugin is available.  I'm getting
compilation errors when attempting this because the byte compiler
doesn't know the relevant variables are available.

Surrounded the interesting forms `with-eval-after-load', though the
warning is not actually gone this does at least allow me to include a
`defvar' form to silence the complaints without introducing the symbol
into the global scope unless it's already there.